server/ServerML/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.31.1
tensorflow==2.20.0         
numpy==1.26.4             
pillow==10.4.0
python-multipart==0.0.9   
